fixed bug where deleted song deletes all comments associated with it,…

+++ backend/routes/api/index.js
@@ -184,9 +184,9 @@ router.delete(
       where: { songId },
     });
     // console.log("songsPlaylistRelation", songsPlaylistRelation)
-    if (songsPlaylistRelation)
-      await SongsPlaylist.destroy({ where: { songId } });
+    if (songsPlaylistRelation) await SongsPlaylist.destroy({ where: { songId } });
 
+    await Comment.destroy({where: {songId: song.id}})
     // console.log("SONG ID", songId)
     await Song.destroy({ where: { id: song.id } });
     // res.redirect(`/`)
